backend/requirements.txt
fastapi
uvicorn
langchain-ollama
langchain-community
langchain-huggingface
langchain-chroma
chromadb
langchain-text-splitters
pypdf
python-dotenv
python-multipart
onnxruntime
tokenizers
tqdm
sentence-transformers
huggingface_hub
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-pptx
docx2txt
unstructured
networkx
tf-keras
comtypes
